+++ ruby/lib/gitaly_server/conflicts_service.rb
@@ -1,3 +1,5 @@
+require 'active_support/core_ext/hash/indifferent_access'
+
 module GitalyServer
   class ConflictsService < Gitaly::ConflictsService::Service
     include Utils
@@ -5,7 +7,8 @@ module GitalyServer
     def list_conflict_files(request, call)
       bridge_exceptions do
         begin
-          resolver = conflict_resolver(request, call)
+          repo = Gitlab::Git::Repository.from_gitaly(request.repository, call)
+          resolver = Gitlab::Git::Conflict::Resolver.new(repo, request.our_commit_oid, request.their_commit_oid)
           conflicts = resolver.conflicts
           files = []
           msg_size = 0
@@ -40,13 +43,41 @@ module GitalyServer
       end
     end
 
-    private
+    def resolve_conflicts(call)
+      bridge_exceptions do
+        header = nil
+        files_json = ""
 
-    def conflict_resolver(request, call)
-      repo = Gitlab::Git::Repository.from_gitaly(request.repository, call)
+        call.each_remote_read.each_with_index do |request, index|
+          if index.zero?
+            header = request.header
+          else
+            files_json << request.files_json
+          end
+        end
 
-      Gitlab::Git::Conflict::Resolver.new(repo, request.our_commit_oid, request.their_commit_oid)
+        repo = Gitlab::Git::Repository.from_gitaly(header.repository, call)
+        remote_repo = Gitlab::Git::GitalyRemoteRepository.new(header.target_repository, call)
+        resolver = Gitlab::Git::Conflict::Resolver.new(remote_repo, header.our_commit_oid, header.their_commit_oid)
+        user = Gitlab::Git::User.from_gitaly(header.user)
+        files = JSON.parse(files_json).map(&:with_indifferent_access)
+
+        begin
+          params = {
+            source_branch: header.source_branch,
+            target_branch: header.target_branch,
+            commit_message: header.commit_message.dup
+          }
+          resolver.resolve_conflicts(repo, user, files, params)
+
+          Gitaly::ResolveConflictsResponse.new
+        rescue Gitlab::Git::Conflict::Resolver::ResolutionError => e
+          Gitaly::ResolveConflictsResponse.new(resolution_error: e.message)
+        end
+      end
     end
+
+    private
 
     def conflict_file_header(file)
       Gitaly::ConflictFileHeader.new(